Building (new) test environment.
