+++ amplify/backend/function/query/function-parameters.json
@@ -3,6 +3,7 @@
   "lambdaLayers": [],
   "secretNames": [
     "TWITTER_BEARER_TOKEN",
-    "DB_PASSWORD"
+    "DB_PASSWORD",
+    "twitter_bearer_token"
   ]
 }
