+++ amplify/backend/function/query/query-cloudformation-template.json
@@ -109,6 +109,26 @@
                   "DB_PASSWORD"
                 ]
               ]
+            },
+            "twitter_bearer_token": {
+              "Fn::Join": [
+                "",
+                [
+                  {
+                    "Fn::Sub": [
+                      "/amplify/${appId}/${env}/AMPLIFY_${functionName}_",
+                      {
+                        "appId": "dtmxl3q3i7oix",
+                        "env": {
+                          "Ref": "env"
+                        },
+                        "functionName": "query"
+                      }
+                    ]
+                  },
+                  "twitter_bearer_token"
+                ]
+              ]
             }
           }
         },
