+++ angular.json
@@ -105,11 +105,11 @@
                 }
               ]
             },
-            "pwsdemo": {
+            "test": {
               "fileReplacements": [
                 {
                   "replace": "src/environments/environment.ts",
-                  "with": "src/environments/environment.pwsdemo.ts"
+                  "with": "src/environments/environment.test.ts"
                 }
               ]
             }
